Exercise 4.17 - 4.23: Added token-based user authentication and updated tests.

+++ controllers/blogs.js
@@ -1,8 +1,11 @@
+const jwt = require("jsonwebtoken")
 const blogsRouter = require("express").Router()
 const Blog = require("../models/blog")
+const User = require("../models/user")
+
 
 blogsRouter.get("/", async (request, response) => {
-  const blogs = await Blog.find({})
+  const blogs = await Blog.find({}).populate("user", { username: 1, name: 1 } )
   response.json(blogs)
 })
 
@@ -17,12 +20,38 @@ blogsRouter.get("/:id", async (request, response) => {
 })
 
 blogsRouter.post("/", async (request, response) => {
-  const blog = new Blog(request.body)
+  const body = request.body
+  const user = request.user
+
+  const blog = new Blog({
+    title: body.title,
+    author: body.author,
+    url: body.url,
+    likes: body.likes || 0,
+    user: user.id
+  })
+
   const savedBlog = await blog.save()
+  user.blogs = user.blogs.concat(savedBlog._id)
+  await user.save()
   response.status(201).json(savedBlog)
 })
 
 blogsRouter.delete("/:id", async (request, response) => {
+  const user = request.user
+
+  const blog  = await Blog.findById(request.params.id)
+
+  //Check if the blog belongs to the logged in user
+  if (blog.user.toString() !== user.id) {
+    return response.status(401).json({ error: "You don't have the authorization to delete this blog entry." })
+  }
+
+  //Update the user's blog array
+  user.blogs = user.blogs.filter(blog => blog.toString() !== request.params.id)
+  await user.save()
+
+  // Delete blog
   await Blog.findByIdAndDelete(request.params.id)
   response.status(204).end()
 })
